Don't pointlessly cast sizeof's result to size_t

+++ rts/Linker.c
@@ -3645,7 +3645,7 @@ ocGetNames_PEi386 ( ObjectCode* oc )
       }
 
       if (kind != SECTIONKIND_OTHER && end >= start) {
-          if ((((size_t)(start)) % (size_t)sizeof(void *)) != 0) {
+          if ((((size_t)(start)) % sizeof(void *)) != 0) {
               barf("Misaligned section: %p", start);
           }
